Reorganize frontend components into pages/ and elements/ sub-folders (issue #487) (#504)
