+++ frontend/package.json
@@ -34,6 +34,7 @@
     "@vitejs/plugin-react": "^5.1.1",
     "babel-plugin-react-compiler": "^1.0.0",
     "c8": "11.0.0",
+    "esbuild": "^0.28.0",
     "eslint": "^9.39.1",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-complexity": "^1.0.2",
