+++ frontend/spec/components/Jobs_spec.js
@@ -2,7 +2,7 @@ import { createElement } from 'react';
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { MemoryRouter, Route, Routes } from 'react-router-dom';
-import Jobs from '../../src/components/Jobs.jsx';
+import Jobs from '../../src/components/pages/Jobs.jsx';
 import noop from '../../src/utils/noop.js';
 
 const flushAsync = () => act(async () => { await new Promise((r) => setTimeout(r, 0)); });
