+++ frontend/spec/components/Layout_spec.js
@@ -2,7 +2,7 @@ import { createElement } from 'react';
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { MemoryRouter, Routes, Route } from 'react-router-dom';
-import Layout from '../../src/components/Layout.jsx';
+import Layout from '../../src/components/pages/Layout.jsx';
 import noop from '../../src/utils/noop.js';
 
 const flushAsync = () => act(async () => { await new Promise((r) => setTimeout(r, 0)); });
