+++ frontend/spec/components/LogsPageHelper_spec.js
@@ -1,6 +1,6 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
-import LogsPageHelper from '../../src/components/helpers/LogsPageHelper.jsx';
+import LogsPageHelper from '../../src/components/pages/helpers/LogsPageHelper.jsx';
 
 describe('LogsPageHelper', () => {
   describe('.build', () => {
