+++ frontend/spec/components/LogsPage_spec.js
@@ -1,7 +1,7 @@
 import { createElement } from 'react';
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
-import LogsPage from '../../src/components/LogsPage.jsx';
+import LogsPage from '../../src/components/pages/LogsPage.jsx';
 import noop from '../../src/utils/noop.js';
 
 const flushAsync = () => act(async () => { await new Promise((r) => setTimeout(r, 0)); });
