+++ frontend/spec/components/ReadyCountdown_spec.js
@@ -1,7 +1,7 @@
 import { createElement } from 'react';
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
-import ReadyCountdown from '../../src/components/ReadyCountdown.jsx';
+import ReadyCountdown from '../../src/components/elements/ReadyCountdown.jsx';
 
 const flushAsync = () => act(async () => { await new Promise((r) => setTimeout(r, 0)); });
