+++ frontend/spec/components/StatsHeader_spec.js
@@ -2,7 +2,7 @@ import { createElement } from 'react';
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { MemoryRouter } from 'react-router-dom';
-import StatsHeader from '../../src/components/StatsHeader.jsx';
+import StatsHeader from '../../src/components/elements/StatsHeader.jsx';
 import noop from '../../src/utils/noop.js';
 
 const flushAsync = () => act(async () => { await new Promise((r) => setTimeout(r, 0)); });
