--- frontend/src/components/LogsPage.jsx
@@ -1,8 +1,0 @@
-import Logs from './Logs.jsx';
-import fetchLogs from '../clients/LogsClient.js';
-
-function LogsPage() {
-  return <Logs fetchLogs={fetchLogs} />;
-}
-
-export default LogsPage;
