+++ frontend/src/components/elements/CollapsibleSection.jsx
