+++ frontend/src/components/elements/JobDetails.jsx
