+++ frontend/src/components/elements/JobStatItem.jsx
