+++ frontend/src/components/elements/StatsHeader.jsx
