+++ frontend/src/components/elements/helpers/LogsHelper.jsx
