+++ frontend/src/components/elements/helpers/StatsHeaderHelper.jsx
