+++ frontend/src/components/pages/LogsPage.css
