+++ frontend/src/components/pages/helpers/JobHelper.jsx
