+++ frontend/src/components/pages/helpers/LogsPageHelper.jsx
